refactor theta hash table into raw table
